Support setting groups / use double instead of float

+++ BrickController2/BrickController2/Protocols/BuWizz3Protocol.cs
@@ -2,11 +2,11 @@
 
 internal static class BuWizz3Protocol
 {
-    public const float CurrentLimitStep = 35f;
+    public const double CurrentLimitStep = 35;
 
     public static byte[] ActivteShelfMode() => [ 0xA1 ];
 
-    public static byte[] SetCurrentLimits(float v0, float v1, float v2, float v3, float v4, float v5)
+    public static byte[] SetCurrentLimits(double v0, double v1, double v2, double v3, double v4, double v5)
     {
         byte l0 = ConvertToRaw(v0);
         byte l1 = ConvertToRaw(v1);
@@ -16,6 +16,7 @@ internal static class BuWizz3Protocol
         byte lB = ConvertToRaw(v5);
         return [0x38, l0, l1, l2, l3, lA, lB];
 
-        static byte ConvertToRaw(float currentLimit) => (byte)(currentLimit / CurrentLimitStep);
+        static byte ConvertToRaw(double currentLimit)
+            => (byte)(currentLimit / CurrentLimitStep);
     }
 }
